HDPI-8390:Moving caseworker events & pages into caseworker folder
