+++ src/main/java/uk/gov/hmcts/reform/pcs/ccd/event/caseworker/changecasestate/ChangeCaseState.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.pcs.ccd.event.changecasestate;
+package uk.gov.hmcts.reform.pcs.ccd.event.caseworker.changecasestate;
 
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Component;
@@ -15,7 +15,7 @@ import uk.gov.hmcts.reform.pcs.ccd.accesscontrol.UserRole;
 import uk.gov.hmcts.reform.pcs.ccd.common.PageBuilder;
 import uk.gov.hmcts.reform.pcs.ccd.domain.PCSCase;
 import uk.gov.hmcts.reform.pcs.ccd.domain.State;
-import uk.gov.hmcts.reform.pcs.ccd.page.changecasestate.ChangeCaseStatePage;
+import uk.gov.hmcts.reform.pcs.ccd.page.caseworker.changecasestate.ChangeCaseStatePage;
 import uk.gov.hmcts.reform.pcs.ccd.util.AddressFormatter;
 
 import static uk.gov.hmcts.reform.pcs.ccd.accesscontrol.CaseworkerRoles.CASEWORKER_ROLES;
