+++ src/main/java/uk/gov/hmcts/reform/pcs/ccd/page/caseworker/documentamend/AmendDocumentDetailsPage.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.pcs.ccd.page.documentamend;
+package uk.gov.hmcts.reform.pcs.ccd.page.caseworker.documentamend;
 
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Component;
