+++ src/test/java/uk/gov/hmcts/reform/pcs/ccd/page/caseworker/uploaddocument/UploadADocumentTest.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.pcs.ccd.page.caseworkeruploaddocument;
+package uk.gov.hmcts.reform.pcs.ccd.page.caseworker.uploaddocument;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
